Add a method to generate histogram lines
For each word in an input paragraph, we want to print out a line displaying the word, a pipe character, an equals sign for each occurrence of the word and the number of occurrences. This commit adds a Histogram class with a static generateHistogramLine method to do just that.

Passed a string (the word) and a number (of occurrences), the method returns a string matching the spec for that word's number of occurrences.

This commit includes a basic JUnit test for the method as well.