Add code to count the occurrences of words in a string
I chose a Map structure to represent the word counts in this project. The bulk of this commit contains a static WordCount class method to generate Map<String, Integer> from a provided string.

The method strips punctuation from the string, then splits it on white space. For each non-empty string in the resulting split array, the method either adds it to the map with a value of 1 (for one occurrence) or adds one to the value if it already exists in the map. The commit contains tests for edge cases such as capitalization, words with apostrophes, and white space separating words.

Finally, this commit also adds a constructor for the WordCount class itself. All the constructor does is take a given string and store it as a word count map, which can be reached with a public getter method.

I decided to split the word count creation into its own static method rather than putting it all in the constructor for a few reasons:

1) Should we want to change how we create the word count maps (i.e. maybe we DO want to treat capitalization as a different word), we'd just have to cook up a new method and change one line in the class constructor to use it instead.

2) I've found it's a bit more cumbersome to test constructor logic, as each unit test has to concern itself with calling the constructor and then asserting using getter methods.