Add a generic method to sort a Map<T, Integer> by its integer values, highest first
The simplest way I could think of to create a histogram is a line at a time, using the Histogram class's generateHistogramLine() method. However, this requires passing the counted words to the method in sorted order, as we want more frequent words on top of the histogram.

This commit adds a sortMapByIntegerValueHighestFirst() method to the WordCount class. The method takes a Map<T, Integer> and constructs and returns a LinkedHashMap containing all the same entries inserted in order from highest value first. Iterating over this Map (i.e. adding entries to a list, as is done in a new test for the method) guarantees the list will always be populated in that same highest-value-first order.

One potential refactor would be to pass a flag to the sorting method telling it whether we wanted to sort values highest-to-lowest or vice versa, to create a more generic and reusable sorting function.