Complete application to build word count histogram from input paragraph
This commit adds a writeHistogramToFile method to the top-level App class. Given String parameters for input and output paths, the method builds a word count for the input file, sorts it by highest occurring words and passes each word and occurrence count to the Histogram function to generate a histogram line. The method then writes that master String of histogram lines to a file it creates (or overwrites, if it already exists) at the given output path.

This commit includes a test for that method, which uses a separate text file in the test directory and asserts the result of running the method on it and creating a new file in that same directory.

The commit modifies the App class's main method to call the writeHistogramToFile method. Should a user prefer, this commit has updated the Gradle configuration to let them pass custom input and output file parameters. Without them, the program defaults to reading from 'input.txt' and writing to 'output.txt', both in the project's root directory.

Finally, the commit git ignores the output results for both the test case and the live main method.